Internal

This chapter documents the functions that can be specially defined for own result data structures.

Functions	Description

hasSignal	Returns true if signal is available in result

getRawSignal	Returns the raw signal from result

getNames	Return a vector of the names that are present in result

getDefaultHeading	Return default heading

getSignalInfo	Return information about a signal

Access Result

ModiaPlot.hasSignal — Function

hasSignal(result, name)

Returns true if signal name is available in result.

ModiaPlot.getRawSignal — Function

(isConstant, signal) = getRawSignal(result, name)

Returns result time series signal of name (an error is raised, if name is not known).

If isConstant=false, then signal[i] is the value of the signal at time instant i.

If isConstant=true, then signal is the value of the signal and this value holds for all time instants.

typeof(value) must be either <:Number or <:AbstractArray with eltype(value) <: Number.

ModiaPlot.getNames — Function

getNames(result)

Return a vector of the names that are present in result.

ModiaPlot.getDefaultHeading — Function

getDefaultHeading(result)

Return default heading.

ModiaPlot.getSignalInfo — Function

(isConstant, sigSize, sigElType, sigUnit) = getSignalInfo(result, name)

Return information about a signal, given the name of the signal:

isConstant: = true, if signal is constant.

sigSize: size(signal)

sigElType: Element type of signal without unit

sigUnit: Unit of signal